# specs/001-linux-host-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: IC Design Environment Guard API
  version: 0.1.0-mvp
  description: >
    Contract for the MVP local host agent HTTP API. All privileged routes require
    the generated local bearer token. Metrics remote access is governed by
    network allowlist rather than user login credentials.
servers:
  - url: http://127.0.0.1:8765
    description: Default local-only agent endpoint
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: Exchange generated local bearer token for an authenticated session context
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 1
                  description: Generated local bearer token; never logged or echoed.
      responses:
        '200':
          description: Login accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      summary: End the current authenticated session context
      responses:
        '204':
          description: Logout accepted
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/services/{service_id}/start:
    post:
      summary: Start a configured service if allowed
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '202':
          description: Operation accepted or already satisfied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOperationResponse'
        '400':
          description: Operation not allowed for configured service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/services/{service_id}/stop:
    post:
      summary: Stop a configured service if allowed
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '202':
          description: Operation accepted or already satisfied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOperationResponse'
        '400':
          description: Operation not allowed for configured service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/terminals:
    get:
      summary: List terminal sessions owned by the local administrator
      responses:
        '200':
          description: Terminal sessions
          content:
            application/json:
              schema:
                type: object
                required: [terminals]
                properties:
                  terminals:
                    type: array
                    items:
                      $ref: '#/components/schemas/TerminalSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a new terminal session
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTerminalRequest'
      responses:
        '201':
          description: Terminal session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: generated-local-token
  parameters:
    ServiceId:
      name: service_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_.-]+$'
...
  responses:
    Unauthorized:
      description: Missing or invalid generated local bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        correlation_id:
          type: string
    LoginResponse:
      type: object
      required: [actor, token_type]
      properties:
        actor:
          type: string
        token_type:
          const: bearer
...
    ServiceOperationResponse:
      type: object
      required: [operation_id, service_id, operation, result]
      properties:
        operation_id:
          type: string
        service_id:
          type: string
        operation:
          enum: [start, stop, restart, status, healthcheck]
        result:
          enum: [success, already_in_state, rejected, timeout, failed]
        failure_reason:
          type: string
          nullable: true
...
    CreateTerminalRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 80
        rows:
          type: integer
          minimum: 1
        cols:
          type: integer
          minimum: 1
        cwd:
          type: string
    TerminalSession:
      type: object
      required: [id, owner, title, status, output_cursor, replay_buffer_start_cursor, created_at, last_active_at, idle_timeout_minutes]
      properties:
        id:
          type: string
        owner:
          type: string
        title:
          type: string
        pid:
          type: integer
          nullable: true
        rows:
          type: integer
        cols:
          type: integer
        status:
          enum: [creating, running, exited, closed, timed_out, failed, unknown]
        output_cursor:
          type: integer
          minimum: 0
        replay_buffer_start_cursor:
          type: integer
          minimum: 0
        idle_timeout_minutes:
          type: integer
          minimum: 30
          maximum: 120
          default: 60
        created_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
        exited_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        close_reason:
          type: string
          nullable: true
...